Added create Interval and Workout implementation
Added most of the functionality to the CreateWorkout and finished the CreateInterval. Also attempted multiple ways to present a modal version of the PickerView, however this appears to be completely deprecated and replaced with UIAlertController. This new controller does allow for the addition of subviews, but I was not able to get it to be placed correctly. Most likely need to be able to do some sort of autolayout programmatically within the view container.